pyproject.toml
[project]
name = "repomap"
version = "0.1.0"
description = "Deterministic codebase intelligence for AI coding agents"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.9.0",
    "networkx>=3.2",
    "tree-sitter>=0.24.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
repomap = "repomap.server:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
